// api/shopify/types.ts
export type ProductCollectionFilter =
	| { available: boolean }
	| { productType: string }
	| { productVendor: string }
	| {
			variantOption: {
				name: string
				value: string
			}
	  }
	| {
			productMetafield:
				| {
						namespace: string
						key: string
						value: string
				  }
				| {
						price: {
							min: number
							max?: number
						}
				  }
	  }
	| { tag: string }

export type ProductCollectionFilters = ProductCollectionFilter[]

export type ProductCollectionSortKey =
	| 'BEST_SELLING'
	| 'COLLECTION_DEFAULT'
	| 'CREATED'
	| 'ID'
	| 'MANUAL'
	| 'PRICE'
	| 'RELEVANCE'
	| 'TITLE'

export type ShopifyQueryParams = {
	query?: string
	first?: number
	sortKey?: ProductCollectionSortKey
	reverse?: boolean
	after?: string
}

export type QueryResponse = {
	meta?: {
		hasNextPage: boolean
		endCursor: string
	}
	data: any
	error?: any
}
// ...
export type Address = {
	id?: string
	phone?: string
	firstName: string
	lastName: string
	address1: string
	address2: string
	city: string
	province: string
	country: string
	zip: string
}

export type DefaultAddress = Address & {
	name: string
}

export type Customer = {
	id: string
	displayName: string
	email: string
	firstName: string
	lastName: string
	phone?: string
	createdAt: string
	updatedAt: string
	acceptsMarketing: boolean
	addresses: Address[]
	defaultAddress?: DefaultAddress
	lastIncompleteCheckout?: {
		id: string
		createdAt: string
		updatedAt: string
		webUrl: string
	}
	password?: string
	passwordConfirmation?: string
}